requirements.txt

beautifulsoup4==4.12.3

bs4==0.0.2

chardet==5.2.0

click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6

firebase-admin==6.6.0
Flask==3.1.0
flatbuffers==24.3.25
frontend==0.0.3

gdown==5.2.0
bcrypt
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
DateTime
groq==0.12.0
grpcio==1.68.1
secure-smtplib
gunicorn==23.0.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7

httpx==0.27.2
huggingface-hub==0.26.2
itsdangerous==2.2.0

joblib==1.4.2


langchain-core==0.3.28
langchain-groq==0.2.1

language_data==1.3.0
mailtrap==2.0.1

Markdown==3.7

nltk==3.9.1

openai==1.58.1
opencv-python==4.10.0.84

orjson==3.10.12

pandas==2.2.3
pdfplumber==0.11.4
pendulum

protobuf==5.28.3

pydantic==2.10.2
pydantic_core==2.27.1

PyMuPDF==1.24.14
pyngrok==7.2.1

pypdf==5.1.0
pypdfium2==4.30.0


python-dateutil==2.9.0.post0
python-dotenv==1.0.1

gensim
regex==2024.11.6

scikit-learn
flask-cors
sentence-transformers==3.4.1
torch

tabulate==0.9.0
tenacity==9.0.0


uvicorn==0.32.1
waitress==3.0.2
watchdog==6.0.0
websocket-client==1.8.0

